refactor(compiler): use memory cache for language service only
